% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_censor_weighting.R
\name{make_surv_response}
\alias{make_surv_response}
\title{Construct a survival response}
\usage{
make_surv_response(data, follow_up, event)
}
\arguments{
\item{data}{A data frame with follow-up and event columns.}

\item{follow_up}{The name of the follow-up time column.}

\item{event}{The name of the event indicator column.}
}
\value{
An object of class \code{"Surv"}.
}
\description{
Construct a survival response
}
\examples{
data(lungcancer)
surv_response <- make_surv_response(
  lungcancer,
  follow_up = "fup_obs",
  event = "death"
)
}
