% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trial_data.R
\name{read_trial_data}
\alias{read_trial_data}
\title{Read trial-style data from CSV}
\usage{
read_trial_data(file, show_col_types = FALSE)
}
\arguments{
\item{file}{Path to a CSV file.}

\item{show_col_types}{Passed to \code{\link[readr:read_csv]{readr::read_csv()}}.}
}
\value{
A tibble.
}
\description{
Read trial-style data from CSV
}
\examples{
csv_file <- tempfile(fileext = ".csv")
writeLines(
  c(
    "id,follow_up,event,treatment",
    "1,10,1,A",
    "2,12,0,B"
  ),
  con = csv_file
)
trial_data <- read_trial_data(csv_file)
}
